Change CLI tests to work with typer
